[APM] Move style variables to commons
