+++ x-pack/plugins/apm/common/variables.ts
