+++ x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/WaterfallContainer/Waterfall/WaterfallItem.tsx
@@ -17,7 +17,7 @@ import {
   px,
   unit,
   units
-} from '../../../../../../style/variables';
+} from 'x-pack/plugins/apm/common/variables';
 import { IWaterfallItem } from './waterfall_helpers/waterfall_helpers';
 
 type ItemType = 'transaction' | 'span';
