+++ x-pack/plugins/apm/public/components/shared/KueryBar/Typeahead/Suggestions.js
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { isEmpty } from 'lodash';
 import Suggestion from './Suggestion';
-import { units, colors, px, unit } from '../../../../style/variables';
+import { units, colors, px, unit } from 'x-pack/plugins/apm/common/variables';
 import { rgba } from 'polished';
 
 const List = styled.ul`
