+++ x-pack/plugins/apm/public/components/shared/StickyProperties/index.tsx
@@ -17,7 +17,7 @@ import {
   truncate,
   unit,
   units
-} from '../../../style/variables';
+} from 'x-pack/plugins/apm/common/variables';
 
 export interface IStickyProperty {
   val: any;
